Fix overflow bug in get_pin_id for interrupts
Widen GPIO pin parameters from uint16_t to uint32_t for get_pin_id, ch32_interrupt_enable, ch32_interrupt_disable and _gpio_exti_callback to prevent an overflow bug when getting the pin id for pins greater than 15.
Replace duplicated EXTI_Line19 with EXTI_Line18 in the exti_lines table;
Rename gpio_port_souce to gpio_port_source (and update its uses) for clarity, and sync function prototypes in the header. These changes correct an incorrect EXTI mapping, fix a variable-name typo, and ensure consistent types across the API.

+++ cores/arduino/ch32/interrupt.h
@@ -53,8 +53,8 @@
 
 
   /* Exported functions ------------------------------------------------------- */
-  void ch32_interrupt_disable(GPIO_TypeDef *port, uint16_t pin);
-  void ch32_interrupt_enable(GPIO_TypeDef *port, GPIOMode_TypeDef io_mode,uint16_t pin, void (*callback)(void), EXTIMode_TypeDef it_mode, EXTITrigger_TypeDef trigger_mode);
+  void ch32_interrupt_disable(GPIO_TypeDef *port, uint32_t pin);
+  void ch32_interrupt_enable(GPIO_TypeDef *port, GPIOMode_TypeDef io_mode,uint32_t pin, void (*callback)(void), EXTIMode_TypeDef it_mode, EXTITrigger_TypeDef trigger_mode);
 
 
 #endif
